blog/requirements.txt
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
python-dotenv==0.15.0
celery==4.4.1
redis==3.4.1
psycopg2

